+++ Anime4KCore/src/Anime4K.cpp
@@ -687,9 +687,9 @@ void Anime4KCPP::Parameters::reset() noexcept
 {
     passes = 2;
     pushColorCount = 2;
-    strengthColor = 0.3F;
-    strengthGradient = 1.0F;
-    zoomFactor = 2.0F;
+    strengthColor = 0.3;
+    strengthGradient = 1.0;
+    zoomFactor = 2.0;
     fastMode = false;
     videoMode = false;
     preprocessing = false;
@@ -705,9 +705,9 @@ void Anime4KCPP::Parameters::reset() noexcept
 Anime4KCPP::Parameters::Parameters(
     int passes,
     int pushColorCount,
-    float strengthColor,
-    float strengthGradient,
-    float zoomFactor,
+    double strengthColor,
+    double strengthGradient,
+    double zoomFactor,
     bool fastMode,
     bool videoMode,
     bool preprocessing,
